churn_prediction/deployment/requirements.txt
streamlit
pandas
seaborn
matplotlib
plotly
numpy
Pillow
scikit-learn==1.2.2
tensorflow